Core/Guild: Improvements - Fixed the issue with saving and loading guild master permissions from the database; - Refactoring the "Guild" class; - Changed the type of monetary constants;

+++ Source/Framework/Constants/PlayerConst.cs
@@ -72,9 +72,9 @@ namespace Framework.Constants
 
     public struct MoneyConstants
     {
-        public const int Copper = 1;
-        public const int Silver = Copper * 100;
-        public const int Gold = Silver * 100;
+        public const long Copper = 1;
+        public const long Silver = Copper * 100;
+        public const long Gold = Silver * 100;
     }
 
     public enum TradeSlots
